u-boot-imx: build and update bl31 binary for imx8mp-var-dart
Instead of copying the bl31-imx8mp.bin directly, this commit modifies
the u-boot-imx.sh script to first build the bl31 binary for the
imx8mp-var-dart machine using the imx-atf source. The built binary is
then copied to the appropriate location in imx-mkimage.

This ensures the latest bl31 binary is used for the imx8mp-var-dart target.

Signed-off-by: Diego Dorta <diego.d@variscite.com>